# specs/openapi.yaml
openapi: 3.0.3
info:
  title: HouseFlow API
  description: |
    API pour l'application HouseFlow - Suivi d'entretien de maison.

    ## MVP Features
    - Authentification (inscription, connexion, déconnexion)
    - Gestion des maisons (CRUD)
    - Gestion des appareils (CRUD)
    - Gestion des types d'entretien (CRUD)
    - Logging des entretiens effectués
    - Calcul des scores de progression

    ## Conventions
    - **Erreurs** : ProblemDetails (`application/problem+json`, schéma `ErrorResponse`) avec un
      `code` machine quand l'interface doit réagir (voir `ErrorResponse.code`).
    - **403 / 404** : 404 (`not_found`) si la ressource — ou la maison qui la contient — n'existe
      pas ; 403 (`forbidden`) si elle existe mais que l'appelant n'a pas le rôle requis (R5).
    - **Dates calendaires** (date d'entretien, échéance, date d'installation) : un jour du
      calendrier Europe/Paris, transporté comme **minuit UTC de ce jour**
      (`yyyy-MM-ddT00:00:00Z`). Les réponses l'envoient toujours sous cette forme
      (`format: date-time`) ; les requêtes acceptent `yyyy-MM-dd` ou `yyyy-MM-ddT00:00:00Z`
      (`format: date`) — seule la partie date est prise en compte, jamais un instant local.
  version: 1.0.0
  contact:
    name: HouseFlow Team
servers:
  - url: http://localhost:5203/api/v1
    description: Local Development Server
  - url: https://api.houseflow.app/v1
    description: Production Server
...
paths:
...
  /houses/{houseId}/members:
    parameters:
      - $ref: '#/components/parameters/houseId'
    get:
      operationId: getHouseMembers
      summary: Membres d'une maison
      description: Visible par tout membre de la maison (le propriétaire inclus, rôle `Owner`).
      tags: [Members]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Membres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseMember'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /members/{memberId}/role:
    parameters:
      - $ref: '#/components/parameters/memberId'
    put:
      operationId: updateMemberRole
      summary: Changer le rôle d'un membre (propriétaire uniquement)
      tags: [Members]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: Membre modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseMember'
        '400':
          description: Rôle inconnu (seuls les noms `CollaboratorRW`, `CollaboratorRO`, `Tenant` sont acceptés), ou rôle du propriétaire
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /members/{memberId}/permissions:
    parameters:
      - $ref: '#/components/parameters/memberId'
    put:
      operationId: updateMemberPermissions
      summary: Droits fins d'un locataire (propriétaire uniquement)
      tags: [Members]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberPermissionsRequest'
      responses:
        '204':
          description: Droits modifiés
        '400':
          description: Le membre n'est pas locataire
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /members/{memberId}:
    parameters:
      - $ref: '#/components/parameters/memberId'
    delete:
      operationId: removeMember
      summary: Retirer un membre (propriétaire uniquement)
      tags: [Members]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Membre retiré
        '400':
          description: Le propriétaire ne peut pas se retirer lui-même
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /collaborators:
    get:
      operationId: getAllCollaborators
      summary: Collaborateurs de toutes les maisons possédées
      description: |
        Pour chaque maison dont l'appelant est propriétaire : ses membres (hors propriétaire) et
        ses invitations en attente non expirées. Maisons partagées avec l'appelant non incluses.
      tags: [Members]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Collaborateurs par maison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllCollaboratorsResponse'
  /houses/{houseId}/invitations:
    parameters:
      - $ref: '#/components/parameters/houseId'
    post:
      operationId: createInvitation
      summary: Inviter quelqu'un (propriétaire uniquement)
      description: |
        Crée une invitation pour `email` avec `role` (hors `Owner`), valable 7 jours, 20 au
        maximum en attente par maison. **Aucun email n'est envoyé** : le propriétaire copie le
        lien `/invitations/{token}` et le transmet lui-même. L'email est conservé (donnée
        personnelle d'un tiers, voir docs/gdpr/processing-register.md) : il verrouille
        l'inscription par ce lien et est purgé avec l'invitation.
      tags: [Invitations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: Invitation créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400':
          description: |
            Rôle ou email invalide, ou limite de 20 invitations en attente atteinte
            (`invitation_limit_reached`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            `already_member` (cet email est déjà membre de la maison) ou
            `invitation_already_pending` (une invitation est déjà en attente pour cet email)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      operationId: getHouseInvitations
      summary: Invitations en attente (propriétaire uniquement)
      description: |
        Invitations non acceptées, non refusées et non annulées, y compris celles dont le
        délai est dépassé (`isExpired`, « Renvoyer » les relance), les plus récentes d'abord.
      tags: [Invitations]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invitation'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /invitations/{token}:
    parameters:
      - $ref: '#/components/parameters/invitationToken'
    get:
      operationId: getInvitationInfo
      summary: Informations publiques d'une invitation (P04)
      description: |
        Accessible sans authentification. Avec un JWT, `isAlreadyMember` indique si l'appelant
        est déjà membre de la maison (P04 redirige alors vers P09) ; sans JWT il vaut null.
        `email` n'est renvoyé que tant que l'invitation est utilisable (en attente et non
        expirée) ; il vaut null une fois l'invitation acceptée, refusée, annulée ou expirée
        (minimisation, RGPD art. 5-1-c).
      tags: [Invitations]
      security: []
      responses:
        '200':
          description: Invitation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationInfo'
        '404':
          $ref: '#/components/responses/NotFound'
  /invitations/{token}/accept:
    parameters:
      - $ref: '#/components/parameters/invitationToken'
    post:
      operationId: acceptInvitation
      summary: Rejoindre la maison (P04 état B)
      description: |
        Réservé au compte dont l'adresse e-mail est celle de l'invitation (insensible à la casse ;
        les invitations antérieures au champ `email` ne sont pas contrôlées).
      tags: [Invitations]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invitation acceptée (idempotent pour le même utilisateur)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptInvitationResponse'
        '400':
          description: '`invitation_invalid` (expirée, refusée, annulée, déjà utilisée), `own_invitation`, `already_member`, `invitation_email_mismatch` (le compte connecté n''a pas l''adresse e-mail de l''invitation)'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /invitations/{token}/decline:
    parameters:
      - $ref: '#/components/parameters/invitationToken'
    post:
      operationId: declineInvitation
      summary: Refuser une invitation (P04 état B, « Refuser »)
      description: |
        L'invitation passe au statut `Declined` et ne peut plus être utilisée. Comme l'acceptation,
        réservé au compte dont l'adresse e-mail est celle de l'invitation.
      tags: [Invitations]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Invitation refusée
        '400':
          description: '`invitation_invalid`, `own_invitation`, `invitation_email_mismatch`'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /invitations/{invitationId}:
    parameters:
      - $ref: '#/components/parameters/invitationId'
    delete:
      operationId: cancelInvitation
      summary: Annuler une invitation en attente (propriétaire uniquement)
      tags: [Invitations]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Invitation annulée (statut `Revoked`)
        '400':
          description: L'invitation n'est plus en attente (`invitation_invalid`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /invitations/{invitationId}/resend:
    parameters:
      - $ref: '#/components/parameters/invitationId'
    post:
      operationId: resendInvitation
      summary: Renvoyer une invitation (propriétaire uniquement)
      description: |
        Génère un **nouveau token** (l'ancien lien cesse de fonctionner) et repousse
        l'expiration à 7 jours. Aucun email n'est envoyé : le nouveau lien est à copier.
        Relancer une invitation expirée la remet en attente : la limite de 20 invitations en
        attente par maison s'applique, ainsi que l'unicité d'une invitation en attente par email.
      tags: [Invitations]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Invitation relancée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400':
          description: L'invitation a été acceptée, refusée ou annulée (`invitation_invalid`), ou limite de 20 invitations en attente atteinte (`invitation_limit_reached`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Une autre invitation est déjà en attente pour cet email (`invitation_already_pending`)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================================
  # API KEYS (US-300 (#188), US-301 (#189), US-302 (#190))
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via /auth/login ou /auth/register
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Clé API générée depuis les paramètres utilisateur
  parameters:
    houseId:
      name: houseId
      in: path
      required: true
      description: ID de la maison (UUID)
      schema:
        type: string
        format: uuid
...
    memberId:
      name: memberId
      in: path
      required: true
      description: ID de l'adhésion (HouseMember.id, UUID)
      schema:
        type: string
        format: uuid
    invitationId:
      name: invitationId
      in: path
      required: true
      description: ID de l'invitation (UUID)
      schema:
        type: string
        format: uuid
    invitationToken:
      name: token
      in: path
      required: true
      description: Token secret de l'invitation (64 caractères hexadécimaux)
      schema:
        type: string
  responses:
    Forbidden:
      description: La ressource existe mais l'appelant n'a pas le rôle requis (`forbidden`, R5)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource inconnue, ou maison inexistante (`not_found`)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Échec transitoire de la base de données après nouvelles tentatives — réessayer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Erreur au format ProblemDetails (RFC 9457, `application/problem+json`). Le frontend
        branche sur `status` et `code`, jamais sur le texte (`detail` est en anglais, non
        destiné à l'affichage). Les 400 de validation de modèle ajoutent `errors`.
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
        detail:
          type: string
          nullable: true
          description: Message technique (anglais)
        code:
          type: string
          nullable: true
          description: |
            Code machine stable, quand l'interface doit réagir spécifiquement :
            `invalid_credentials`, `account_restricted`, `invalid_refresh_token`, `email_taken`,
            `export_rate_limited`, `invitation_invalid`, `invitation_email_mismatch`,
            `invitation_already_pending`, `invitation_limit_reached`, `already_member`,
            `own_invitation`, `forbidden`, `not_found`.
        errors:
          type: object
          nullable: true
          additionalProperties:
            type: array
            items:
              type: string
          description: Erreurs de validation par champ (400 de validation uniquement)

    # ==========================================================================
    # AUTH
    # ==========================================================================
...
    HouseRole:
      type: string
      enum: [Owner, CollaboratorRW, CollaboratorRO, Tenant]
...
    HouseMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ID de l'adhésion (pour /members/{memberId})
        userId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/HouseRole'
        canLogMaintenance:
          type: boolean
        canViewCosts:
          type: boolean
        createdAt:
          type: string
          format: date-time
          description: Date d'arrivée dans la maison
    UpdateMemberRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          description: CollaboratorRW, CollaboratorRO ou Tenant
    UpdateMemberPermissionsRequest:
      type: object
      properties:
        canLogMaintenance:
          type: boolean
          nullable: true
        canViewCosts:
          type: boolean
          nullable: true
    CreateInvitationRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        role:
          type: string
          description: CollaboratorRW, CollaboratorRO ou Tenant
    Invitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        token:
          type: string
          description: Token du lien `/invitations/{token}` (à copier ; aucun email n'est envoyé)
        email:
          type: string
          nullable: true
          description: Email de la personne invitée (null pour les invitations antérieures)
        role:
          $ref: '#/components/schemas/HouseRole'
        status:
          type: string
          enum: [Pending, Accepted, Expired, Revoked, Declined]
        isExpired:
          type: boolean
          description: Délai de 7 jours dépassé (« Renvoyer » relance l'invitation)
        houseId:
          type: string
          format: uuid
        houseName:
          type: string
        createdByName:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    InvitationInfo:
      type: object
      description: Vue publique d'une invitation (P04)
      properties:
        id:
          type: string
          format: uuid
        houseId:
          type: string
          format: uuid
        houseName:
          type: string
        role:
          $ref: '#/components/schemas/HouseRole'
        invitedByName:
          type: string
          description: Prénom et nom de la personne qui invite
        email:
          type: string
          nullable: true
          description: |
            Email invité — verrouille le champ email de l'inscription (P03). Renvoyé uniquement
            tant que l'invitation est utilisable ; null sinon (et pour les invitations antérieures).
        status:
          type: string
          enum: [Pending, Accepted, Expired, Revoked, Declined]
        expiresAt:
          type: string
          format: date-time
        isExpired:
          type: boolean
          description: true si l'invitation n'est plus utilisable (expirée, acceptée, refusée ou annulée) — état C de P04
        isAlreadyMember:
          type: boolean
          nullable: true
          description: Appel authentifié uniquement — l'appelant est déjà membre (P04 → P09). null sans JWT.
    HouseCollaborators:
      type: object
      properties:
        houseId:
          type: string
          format: uuid
        houseName:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/HouseMember'
        pendingInvitations:
          type: array
          items:
            $ref: '#/components/schemas/Invitation'
    AllCollaboratorsResponse:
      type: object
      properties:
        houses:
          type: array
          items:
            $ref: '#/components/schemas/HouseCollaborators'
    AcceptInvitationResponse:
      type: object
      properties:
        houseId:
          type: string
          format: uuid
        houseName:
          type: string
        role:
          $ref: '#/components/schemas/HouseRole'

    # ==========================================================================
    # API Keys (US-300 (#188), US-301 (#189), US-302 (#190))
    # ==========================================================================
...